домашнее задание номер 3
